[FTD][CfToHandshake] Implemented merges between backward producer-consumer
With the commit, the conversion pass is able to add some merges between
each pair of consumer/producer relationship (backward direction).
The logic to suppress and generate the tokens according to the
appropriate conditions is still missing.
Most of the code was taken from `8fd166e1162dcbd6397d4e61107ff2a404db3fe5`.