[CfToHandshake] Fix OneToOneConversion for FTD purposes
The current version of OneToOneConversion was intended for operations
with results used only later in the IR. For some internal MLIR mechanisms,
it was not possible to map the results into values which were used before the
operation itself in the IR. An explicit usage of the `replace` method seems
to solve the issue. This siutation arises only in the FTD methodology, as normal
arithmetic operation do not lead to backward edges.