[HandshakeHoistExtInstances] External can only be called once (#144)
This commit modifies the behavior of the `HandshakeHoistExtInstances`
pass, which now only allows each external function to be instantiated at
most once. We impose this new restriction because we currently have no
clean way of generating multiple set of IO port names for multiple
instantiations of the same external function. With this restriction, the
top-level ports associated to the instantiation of each external function
will now match the latter's argument names as specified in the source
code.

This also adds  a new benchmark to the integration tests called
`external_integration` that can be used to test this functionality.

Unrelated change: this fixes a typo on the Verilog implementation of the
OEHB.