[DOT] `DOTGraph` type and DOT language parser
This introduces a `DOTGraph` type that matches the representation of
"hierarchical" graphs in Graphviz. In addition, this adds a parser for a
subset of the DOT language that creates an instance of this new type.

At this point this is completely disconnected from the DOT printing
logic but eventually, instead of directly converting Handshake functions
to their equivalent in the DOT language, we should convert them to
instances of the new `DOTGraph` type and simply call a `print` method on
the graph that can reliably output a DOT-formatted version of the graph.