[visual-dataflow] Fix visualizer and improve reliability
The dataflow visualizer broke during the past few months due to the big
systematic changes that were merged on the project (e.g., type system,
circuit interface, etc.). This commit fixes it, cleans up the
implementation, and subsantially improves the reliability of parsing the
input DOT and CSV files used to render the graph. In particular, the
tool now uses a proper DOT parser instead of hacky string manipulations
to derive the graph's structure from the input file.