[Handshake] New interface `MemPortOpInterface`
This adds a new parent interface for the `LoadPortOpInterface` and
`StorePortOpInterface` Handshake interfaces, allowing client code to
match all load/store memory ports through a single interface and
factorizing the children interfaces' methods in the parent interface
in the process. Memory port operation definition is cosmetically
modified as well, relying on an additional helper Tablegen class.