[Handshake] Optimize width of address-carrying channels (#157)
Currently, all our address-carrying channels in the IR and in the final
circuit are hardcoded so that their data signal is 32-bits. Since
Dynamatic enforces the size of all memories to be known at compile-time,
we have the opportunity to thin out these signals to the bare minimum
number of bits necessary to index into each particular memory region.

This commit adds this optimization to the default compilation flow so
that all Dynamatic-generated circuits can benefit from it. Main changes
are detailed below.

- The `--handshake-optimize-bitwidths` flag looses its "legacy" flag
which was previously used to explicitly disable this optimization when
we still depended on legacy Dynamatic's toolchain for RTL export. The
pass's address-optimization-related rewrite patterns are also
fixed/refactored to work with all of our current memory interfaces and
ports. A new unit test is added to check for correct behavior in the
presence of a master/slave memory interface combination.
- The `hls-verifier` looses its "aw32" flag, which was previously used
to force all address-carrying channels to be 32-bit wide during
testbench-generation.
- Compilation and simulation scripts are updated to reflect the above
changes.

Fixes #155.