[export-rtl] Support for channels with extra signals
The entire compilation flow but our RTL export tool supports extra
signals; the RTL exporter currently happily ignore such signals when
encountering a dataflow channel in the input IR.

This commit adds full support for extra signals to the tool. Thanks to
the genericity of the exporter, any combination of extra signals an any
channel in the IR should be handled properly. I also took this
opportunity to refactor some of the tool's code, in particular I
factored out very similar parts of the implementation for writing VHDL
and Verilog into common writers to avoid some code duplication.