[export-dot] Argument to control node labels
The `export-dot` currently uses a "pretty version" of each operation
type as a node label in the DOT representation of the Handshake
function. While this is nice for quick visualization purposes, it is not
very helpful when trying to debug something through the DOT file, one
would much rather see each operation's unique name as node labels in
those cases.

This commit adds a new "label-type" argument to the `export-dot` tool
that allows to switch between the two node label types outlined above.
By default the label type is set to "type", which corresponds to the
"pretty-printed" node labels. The opt-in alternative to see operations'
unique name instead is "uid", and can be enabled by e.g., modifying the
invokation of the `export-dot` tool in the compilation script. On the
current HEAD this invokation is at

https://github.com/EPFL-LAP/dynamatic/blob/997cc9cbcdbe985c3b5e012a1346b77289c898d8/tools/dynamatic/scripts/compile.sh#L56-L57

and can be replaced with

```sh
"$DYNAMATIC_EXPORT_DOT_BIN" "$f_handshake" --edge-style=spline \
  --label-type=uname > "$f_dot"
```

to print unique operation names as labels.

Closes #178.