[Analysis] Control Dependence Analysis (#186)
The control dependence analysis is a further step into #177. 
The analysis answers the following question: "Given a block `bb` in the
CFG, which are the blocks whose control decisions impact the execution
of `bb`"?
The analysis can be applied on an IR with an underlying CFG structure.

The implementation comes directly from its definition in the following
paper:
_Jeanne Ferrante, Karl J. Ottenstein, and Joe D. Warren. 1987. The
program dependence graph and its use in optimization. ACM Trans.
Program. Lang. Syst. 9, 3 (July 1987), 319–349.
https://doi.org/10.1145/24039.24041_