[Handshake] Unify MC/LSQ memory ports  (#193)
The operations representing load and store ports to memory interfaces
are currently different based on the nature of the memory interface they
connect to. We have two types of memory interfaces (MC and LSQ) so that
is a total of 4 ports (2 load ports and 2 store ports). This design
choice was made relatively early in the project's development was and
motivated by the fact that these ports' respective RTL implementation
were different based on whether they connected to an MC or LSQ. At the
Handshake level, however, both types of load ports have the same
interface and behave exactly in the same way; the same applies to both
types of store ports. The differentiation causes some code duplication
and complicate IR transformations of the memory network.

This unifies the `handshake::MCLoadOp` and `handshake::LSQLoadOp`
operations into a single `handshake::LoadOp` operation, as well as the
`handshake::MCStoreOp` and `handshake::LSQStoreOp` operations into a
single `handshake::StoreOp` operation. Notable changes throughout the
codebase include the following.

- Removal of `handshake::LoadOpInterface` and
`handshake::StoreOpInterface`, which were used to factor out common
behavior between the two access ports for each access type. This is
obviously no longer needed.
- Significant simplifications in the `handshake::MemoryPort` class
hierarchy and in the `HandshakeReplaceMemoryInterfaces` pass.
- Introduction of a `handshake::findMemInterface` function that runs a
restricted BFS from a `Value` to identify the potential memory interface
that it "connects" to. This is used in places where we used to use the
specific type of a memory port to identify the specific memory interface
type it connected to.
- Unification of timing models and RTL components. The implementation of
RTL ports for the MC becomes the universal ones for all memory interface
types. This is perhaps undesirable in the long term, but works for now
as these are the most "conservative" ones (the load port has buffers on
channels connecting to the memory interface and the store port joins its
address and data inputs). RTL configurations and timing models are
adjusted as needed.
- There is a special case in the creation of throughput constraints
during smart buffer placement wherein channels to store ports are
omitted from the throughput equations. This was ported directly from the
legacy implementation but does not seem to make much sense. I added a
TODO there as part of this commit for future reference.

This change is related to the discussion in #185. It also is related to
#135/#136, but does not solve the issue (interestingly the benchmark no
longer deadlocks but produces an incorrect result). This requires more
investigation.