Automated LSQ Sizing (#132)
This PR adds a pass, which sizes the Load and Store Queues, to the
minimal depth, which still allows optimal throughput.
This is done by analyzing the timing of LSQ loads/stores for each CFDFC
separately and using the maximum queue sizes of any CFDFC. The pass is
run between the handshake-canonicalization and the handshake-hw
conversion.

### Important side notes
1. There were wrong latencies for some components in the components.json
file, I fixed them to match the latencies I observed in the simulation.
2. Currently this algorithm only works with *fpga20-legacy* buffer
placement, since in *fpl22* and simple-buffer placement, not all the
information is passed through the attributes. When not all necessary
information is available, and therefore LSQ sizing cannot be done, the
default depth will be assigned to the LSQs.

## Overview
The following things were implemented:
1. The pass receives information on CFDFCs and their respective
throughputs from the buffer-placement pass via Handshake attributes (see
#105 ).
1. With this information the CFDFCs are reconstructed and the sizing
algorithm determines the queue depths according to the following paper:
[ Load-Store Queue Sizing for
Efficient Dataflow
Circuits](https://dynamo.ethz.ch/wp-content/uploads/sites/22/2023/01/Load-Store_Queue_Sizing_for_Efficient_Dataflow_Circuits.pdf)
1. The sizing information is passed on as a Handshake attribute in the
corresponding LSQ operation. The attribute is a tuple, with first the
Load-Queue size and second the Store-Queue size. e.g. ``` lsqDepth =
#handshake<lsqDepth[3, 5]> ```
1. Furthermore the logic to read this information is added in the legacy
backend and the new backend to pass it on to the LSQ generator. For
this, the ```LSQGenerationInfo``` class has been modified to read the
```lsqDepth``` attribute. If the attribute does not exist, the default
size (16) will be used for both the load and store queue.

## Arguments
The pass needs the compontents.json file as a timingDB, which is passed
similarly to buffer placement. Furthermore, it is possible to specify
different collision scenarios to the pass, by the
```collisions=none/half/full``` argument. These assume one of the
following cases.
1. *none*: None of the LSQ memory operations have collisions
1. *half*: Every other LSQ memory operation has collisions
1. *full* All of the LSQ memory operations have collisions

## Details
### Algorithm data structure
The sizing algorithm works on the different CFDFCs of the kernel,
determined in buffer placement. The CFDFCs are represented, as an
adjacency list (AdjListGraph made of AdjListNodes), since most of the
algorithm steps are variations of path-finding between different nodes.


## Detailed steps of the algorithm
The sizing algorithm inside the LSQ Sizing pass has the following steps:
1. Reconstuct the CFDFCs
1. Identifying Start Node and Phi Nodes
1. Adding additional edges to account for some edge cases
1. Find the earliest start time of each node and save it 
1. Finding the worst-case II for each CFDFC
1. Set correct IIs for different collision scenarios for each CFDFC
1. Determine allocation times for each CFDFC
1. Determine deallocation times for each CFDFC
1. Calculate queue depths based on all the timing information and IIs
from all CFDFCs

The implementation details are described in the semester thesis report
(*Automated Load-Store Queue Sizing in Dynamatic based on MLIR*)

## Results
The results match the one of the paper, namely the algorithm can find
the optimal LSQ size for the steady-state, given that the correct
sequence of IIs is provided, either by buffer placement or manually. The
detailed results can be found in the semester thesis report: (*Automated
Load-Store Queue Sizing in Dynamatic based on MLIR*)

---------

Co-authored-by: Jiantao Liu <jianliu@ethz.ch>
Co-authored-by: Jiahui Xu <jxu@ethz.ch>