[Speculation] New routeCommitCtrl algorithm (#181)
(Emmet helped me update this comment to the latest one, thanks!)

When the path from speculator to a commit contains a branch, we must add
this complicated structure to route the control channel (orange).
The details of this structure are not super important for this pull
request, only that we must construct it once for each branch on the
path.


![image](https://github.com/user-attachments/assets/de7334a6-6a10-41a2-b55d-ae911e490dfa)

## Change 1: Removal of unnecessary replications.

Before this pull request, traversal began at the speculator, and
branches were replicated as they were encountered.
However, some branches encountered were not actually on paths to commit
units, and therefore were replicated for no reason.
With this pull request, branches are added to a data structure as they
are encountered. If the path does ends in a commit, the branches are
replicated as before. If not, we do not perform replication.
Therefore no unnecessary replication occurs.

## Change 2: Termination of traversal at save-commit and speculation.

Speculation begins at the speculator, and ends when a speculative token
reaches either a commit, save-commit or the speculator itself.
Therefore the traversal should not pass through save-commits or the
speculator, as the routed control token would affect an
already-resolved speculation.
This is a correctness fix.

Note that we initially believed the BFS algorithm could resolve the
issues related to the Change 2; however, this turned out to be
incorrect. Consequently, we updated the solution. Therefore, I want you
to replace the BFS algorithm with the DFS algorithm, which stops at
save-commits or the speculator in the discussion below.