[Experimental][FTD] Improved support library (#207)
As a first step towards the merge of the whole Fast Token Delivery
algorithm, this PR extends what was already included in #187 for the FTD
library.

The main addition is the class `ftd::BlockIndexing`, whose role is to
provide an index to each block so that, if *Bi* dominates *Bj*, then *i
< j*. This is relevant many times in the context of FTD. Some later PRs
will introduce the FTD code directly.

Also, the Control Dependence Analysis library has been slightly refined,
so that it can be also built out of an `mlir::Region`. Again, this will
be convenient for our scopes.