[Experimental][GSAAnalysis] Updated GSA Analysis Structure
The main update to the analysis is the possibility of
running it on an instance of `handshake::MergeOp`.

The GSA analysis aims to be a method to get a phi node in
the SSA representation and obtain its corresponding GSA
equivalent. While in `cf` IR all the initial phi nodes are
provided as block arguments, at a later stage (handshake)
it might be possible to instantiate some merges, working
as additional SSA nodes.
Since the logic to obtain their GSA equivalent is the same,
the choice was to update the analysis pass.