[HDL] Modularizing merges and generalizing their implementation {.v,.vhd}  (#200)
The previous implementation of the `merge` component (all variants of a
`merge` along with a `cmerge`) was not sufficiently general and failed
to function correctly when more than one input received a token in the
same cycle. Specifically, the issue stemmed from the way the `ins_ready`
vector was calculated, which could result in data loss if multiple
inputs received tokens simultaneously.

This problem did not arise in the earlier circuits of Dynamatic
[FPGA'18], as they guaranteed that any merge will receive at most one
token per cycle. However, this guarantee does not hold in general and is
not applicable to our more recent work that supports out-of-order
execution in dataflow circuits [FPGA'24].

Furthermore, the `merge` and `merge_dataless` modules have been
modularized by instantiating `merge_notehb` and `merge_notehb_dataless`,
respectively.

---------

Co-authored-by: Jiahui Xu <48123265+Jiahui17@users.noreply.github.com>

fixes #167