[HDL] Fix verilog export
Commit `ebf54e81` introduces a bug in the component
`merge_notehb_dataless` which does not allow
neither the simulation nor the synthesis.

This commit fixes such issue, relying on the same
techinque used in #211 to fix the synthesis of the
control merge component.

Also, the synthesis script is updated to get rid
of a command which was obsolete with respect to
the current structure of exported HDL files.