[ScfToCf] Fixed comparator optimisation
This commit solves a bug in the switch from a signed to
an unsigned comparator. This modification is employed to
allow a stronger simplifaction in a lower stage of the compiler.

However, depending on how the front-end defines the loop bounds,
it might happend that the upper bound is negative, leading
to uncorrect results if this conditions is not handled.

This makes the benchamrk `symm_float` working without affecting
performances.