[Tools] GitHub Actions workflow (#247)
This branch introduces a GitHub Actions workflow to compile Dynamatic
and run integration tests on every pull request to main.

The workflow runs on a self-hosted runner, since using GitHub's
ephemeral VMs would require downloading all dependencies (one of them
being ModelSim, which is quite heavy) every time the workflow is run. To
ensure security, the GitHub Actions setting "Require approval for all
external contributors" is set. This prevents forked repositories making
pull requests and running arbitrary code on our VM.

Integration tests are run using the `run_integration.py` script in the
repository's root directory. This script is a modified version of the
[script](https://github.com/EPFL-LAP/dynamatic/blob/fpga23-queue-allocation/run_integration.py)
by @pcineverdies, mentioned in #213. Specifically, the original script
was used to compare Fast Token Delivery test results vs original
results, however the modified script only runs Dynamatic without special
features (such as FTD), since it should be as generic as possible.
Additionally, several quality of life features were added (that were
also requested in #213):
- can run only tests that are listed in a chosen text file (instead of
all of them)
- can ignore tests that are listed in a text file
- can set a custom timeout for running tests
- simulates both VHDL and Verilog artifacts
- output shows the simulation time (for Verilog)

The tests take a while to run, but considering the fact that the
workflow is triggered only when making a PR to main, this shouldn't be a
problem.

_Updates:_
With both VHDL and Verilog, the entire workflow takes a bit under 2
hours.
I disabled Verilog now (temporarily) because some of the tests failed
only on Verilog simulation; I will need to look into it. Some of the
errors were:
- `[ERROR COMPARE] Token mismatch: [0x00000000] and [0x00000003] are not
equal (at transaction id 0).`
- `# Loading work.tb_join(arch) # Error loading design` (in report.txt)