[Test] Skip invalid test when building in release mode (#248)
Currently, when building in release mode, the unit test `invalid.mlir`
fails, due to mlir rising a segmentation fault rather than the expected
error message.

On the one hand, we cannot get rid of this test, since it is actually
useful to validate the handshake dialect, and it still runs correctly in
debug mode.
On the other hand, it's important to have a working test suite, to be
integrated in the CI/CD pipeline.

This PR fixes the problem by relying on a CMAKE macro to know whether
the project was built in release or debug mode, and possibly skipping
`invalid.mlir` when generating the test scripts through `ninja`.

I marked this PR as draft as I'd like to get some feedbacks from your
sides regarding the validity of such a mechanism :)

Also closes #164 .