[Refactoring] Removed ExtraSignal::Storage and TypeParameter ExtraSignals (#251)
### Background  

The focus of this pull request is to simplify the memory allocation
logic for `ExtraSignal`s during parsing.
In the original implementation, we used two types during
parsing—`ExtraSignal::Storage` and `ExtraSignal`—to handle extra
signals, as we believed manual memory allocation was necessary at
certain points. However, this approach made the implementation harder to
understand and more prone to memory allocation bugs.

### Summary of Changes

We identified that it was unnecessary to manually manage the storage
type of `ExtraSignal`.
The distinction between `ExtraSignal` and `ExtraSignal::Storage` was the
`name` type (`llvm::StringRef` vs. `std::string`). Initially, we assumed
that the storage class was required to hold `std::string` during
parsing. However, we discovered that `parseKeyword` can accept a
`StringRef*`, meaning it internally handles string allocation.
This eliminates the need for us to allocate memory manually.

Additionally, we addressed another issue regarding the `TypeParameter`
`ExtraSignals` in `HandshakeTypes.td`.
We replaced it with `ArrayRefOfSelfAllocationParameter<ExtraSignal>`.
The allocation here is still necessary to extend the lifetime of the
data beyond its parsing phase. However, it makes more sense to allocate
memory on a per-`ExtraSignal` basis rather than at the array level.

### Impact

These changes simplify the parsing logic and improve the dependency
relationships between Handshake types, making the code much more
interpretable and maintainable.
They also mitigate potential memory allocation issues, such as dangling
pointers, reducing the risk of future bugs.