[Speculation] Removed speculation unit tests (#252)
The speculation unit tests (run with `ninja
check-dynamatic-experimental`) are currently failing.

Initially, I suspected this was due to slight changes in the generated
circuits with speculative units. However, it turns out the root cause is
that the unit tests themselves are unrealistic. For example:

- They lack control tokens.  
- They include multiple condition branches in the same basic block,
which use different conditions.
- Speculative units, like speculator or save-commits (in the case where
we manually place them), are placed in strange places.

These issues have broken some of our assumptions.  

Finally, we've decided to remove unit tests for speculation since
they're not meaningful and are sufficiently covered by simple
integration tests.
In the future, simulation-based unit tests for each speculative unit
could be useful, but that's a separate concern.