[Handshake] Add `handshake::NDWire` operation (#260)
The equivalence checking in #236 requires a non-deteministic wire
operation to correctly decouple the two functions to be tested.

The non-determinstic wire functions as follows:

- When the wire is sleeping it stalls any propagation of tokens. 
- When the wire is running it propagates all tokens.

The wire can switch from sleeping to running at any time. To ensure
correct behaviour under the handshake protocol the state can only go
back to sleeping after a token has been propagated.

For the HDL models the signal `nd_next_state` is internal to the module.
This signal needs to be driven by the formal tool. If the formal tools
does not implicitly treat undriven signals like primary inputs this
needs to be done explicitly.