[export-rtl] PORT_TYPES parameter (#287)
### Background  
The `DATA_TYPE` attribute is commonly used by most operations to pass
type information to the generator or generic RTL file. However, for
complex operations like `MuxOp` or `CMergeOp`, a single-type attribute
may not be sufficient, as it needs to convey type information for
multiple operands.

To address this, we introduce a new parameter called `PORT_TYPES`, which
stores type information for all results and operands in a serialized
JSON string.

### Example  

For a `cmpi` operation:  
```json
{"lhs": "!handshake.channel<i32, [spec: i1]>", "rhs": "!handshake.channel<i32, [spec: i1]>", "result": "!handshake.channel<i1, [spec: i1]>"}
```  

For a `mux` operation (with two inputs):  
```json
{"index": "!handshake.channel<i1>", "ins_0": "!handshake.channel<i32>", "ins_1": "!handshake.channel<i32, [spec: i1]>", "outs": "!handshake.channel<i32, [spec: i1]>"}
```  

### Implementation  

- Added a new function, `RTLMatch::registerPortTypesParameter`, to
register the `PORT_TYPES` parameter by passing the operation.
- Invoked this function in the RTL matching logic before concretizing
the actual RTL.

With this parameter, the `DATA_TYPE` attribute in the HW IR is no longer
needed.