[Equivalence Checking] elastic-miter generator (#259)
This branch is the first step to providing Equivalence Checking Support
in Dynamatic in line with #236. More specifically, it implements the
steps 1 and 2 of the implementation plan. It also contains some of the
rewrite patterns outlined in the paper.

The generator takes two handshake MLIR files as input and outputs a
handshake MLIR, as well as a JSON config file as an output.
The JSON config file will later be used to generate the SMV properties
needed to verify the equivalence.

In the current implementation the number of slots in the decoupling
input/output buffers needs to be provided manually. Later it will be
possible to determine the needed slots automatically.

The elastic-miter generator can be used with following command:

```bash
elastic-miter --lhs=<lhs-file-path> --rhs=<lhs-file-path> -o <out-dir> --bufferSlots=<nr-of-slots>
```

The generation can be tested with:
```bash
build/bin/elastic-miter --lhs=experimental/tools/elastic-miter-generator/rewrites/a_lhs.mlir --rhs=experimental/tools/elastic-miter-generator/rewrit
es/a_rhs.mlir -o experimental/tools/elastic-miter-generator/out/comp --bufferSlots 2
```