[export-rtl] Implement serialize method for RTLTimingType (#292)
The `TIMING` attribute is now passed to the new RTL generator, which is
being developed in #270 (for SMV) and #274 (for VHDL Signal Manager),
particularly when a buffer is generated.

To pass an attribute to a generator, it must be serializable—i.e.,
convertible into a string.

Each attribute (e.g., `TIMING`, `DATA_TYPE`, `FIFO_DEPTH`) has an
associated `RTLType` (Lucas's concept). Other `RTLType`s (such as
`RTLStringType` and `RTLUnsignedType`) already provide a `serialize`
method, but `RTLTimingType`, which is used for the `TIMING` attribute,
did not.

I implemented `serialize` method for `RTLTimingType`. The format is the
same as the assembly format of `TimingAttr`, which can be seen in the
Handshake IR files. I adjusted the format to ensure it can be easily
passed to the Python-based generator- The format follows the convention
used in the `PORT_TYPES` parameter (#287): a Python dictionary is quoted
with `'`.

Serialization example:

```
'#handshake<timing {R: 1}>'
'#handshake<timing {D: 1, V: 1, R: 0}>'
```