[Speculation] Fix Commit Unit Placement on Memory Units (#296)
This PR fixes commit unit placement in memory-related operations.

## Changes

1. Commit units should not be placed before `LoadOp`.
2. Correct commit unit placement in front of `MemoryControllerOp` in the
following way:

![image](https://github.com/user-attachments/assets/f4f99b56-1647-4fa3-a025-c0c6e28306b6)

- Commit units must be placed on the `ctrlEnd` of `MemoryControllerOp`,
which signals that no more requests are incoming.
- In the case of stores, another operand called `ctrl` tracks the number
of store operations. The commit unit should also be placed there.

I didn't consider the case where the load and store operations are
mixed.

## Issues

- Since the current definition of MemoryControllerOp defines most of its
operands as `Variadic<SimpleChannel>`, it is difficult to retrieve the
`ctrl` operand cleanly, resulting in messy code.
- ~~EDIT: I came up with a relatively simpler way. Place commits on
operands of MemoryControllerOp if they are **not** connected to LoadOp
or StoreOp.~~
- EDIT2: Emmet came up with even simpler way: Don't traverse to
MemoryControllerOp from LoadOp

I believe this problem seems to appear throughout the codebase. A
redesign of MemoryControllerOp, LSQOp, and MemPortOp (Load/Store) might
be necessary.

Added a documentation. Please refer to it for further details