[Equivalence Checking] Complete tool (#319)
This branch implements the equivalence checking in dynamatic and
completes #236.

There are some changes to flow, compared to the issue. Everything is now
integrated into a single binary which calls other programs as needed.

The equivalence checking flow takes two MLIR files in the Handshake
dialect, generates an elastic-miter circuit from them, and converts it
into the SMV format. It then creates a wrapper for the generated module,
which includes sequence generators and sequence constraints.
Additionally it creates the required properies needed to formally
verifiy the equivalence of the two circuits. For the actual verification
nuXmv or NuSMV can be used.

Some rewrites, with which the tool can be tested, are provided in
`experimental/test/tools/elastic-miter`. In the same directory the
`prove-rewrites.sh` script runs all tests.

The tool can be used with following signature:
```
elastic-miter --lhs=<lhs-file-path> --rhs=<lhs-file-path> -o <out-dir>
              [--loop=<string>] [--loop_strict=<string>] [--seq_length=<string>] [--token_limit=<string>]
              [--cex]
```

`--lhs=<string>`:
The file containing the left-hand side handshake circuit.

`--rhs=<string>`:
The file containing the left-hand side handshake circuit.

`-o <string>`:
The directory where all intermediate files will be placed.

`--loop=<string>`:
Specify a Loop Condition sequence contraint. Can be used multiple times.
It has the form "\<dataSequence\>,\<controlSequence\>".
The number of tokens in the input with the index dataSequence is
equivalent to the number of false tokens
at the output with the index controlSequence. 
Example:
--loop="0,1"

`--loop_strict=<string>`:
Specify a Strict Loop Condition sequence contraint. Can be used multiple
times.
Works identically to the loop condition sequence contraint, with the
addition that the last token also needs to be false.

`--seq_length=<string>`:
Specify a Sequence Length Relation constraint. Can be used multiple
times.
It controls the relative length of the input sequences. 
The constraint has the form of an arithmetic equation. The number in the
equation will be replaced the respective input with the index of the
number.
Example:
--seq_length="0+1=2" will ensure that the inputs with index 0 and index
1 together produce as many tokens as the input with index 2.

`--token_limit=<string>`:
Specify a Token Limit constraint. Can be used multiple times. 
It has the form "\<inputSequence\>,\<outputSequence\>,\<limit\>".
At any point in time, the number of tokens which are created at the
input with index inputSequence can only be up to "limit" higher than the
number of tokens reaching the output with the index outputSequence.

`--cex`:
Enable counterexamples. In case a property does not pass, this will
generate a counterexample and put it in an XML file.


fixes #309 (moving the test files that are not intended to be used by
dynamatic-opt out)