[Build] Enable lld and ccache for faster compilation of Dynamatic (#347)
After running `cmake --target clean`, I noticed that compiling Dynamatic
was quite slow. The main issue was that `lld`, a much faster linker, was
not being used—`ld` was instead (which is significantly slower). `lld`
is already enabled for building LLVM and Polygeist, and I enabled it for
Dynamatic as well.

(Not sure why it was enabled before I ran `cmake --target clean`—maybe I
had a custom build setup.)

Additionally, I found that `ccache` was not being used for Dynamatic, so
I enabled it to speed up incremental builds.