[SMV] Netlist generator (#297)
We implemented a netlist generator for SMV trying to reuse as much as
possible the current infrastructure as discussed in
https://github.com/EPFL-LAP/dynamatic/issues/203. To do so we add SMV as
an additional HDL to VHDL and SystemVerilog.

The naming convention is the same convention as in MLIR HW: all the
module names and port names are then same as in the HW IR. Some special
names are reserved in SMV (like `T` and `F`), therefore we suggest not
to use them. Future PRs will take into consideration this issue for SMV
and the other HDLs.

There are some differences between VHDL/SV and SMV, that are reflected
in the generator:
- In SMV `clk` and `rst` signals are not needed
- In SMV external files need to be included with preprocessor directives
(`SMVWriter::writeIncludes`)
- SMV doesn't need internal signals to connect the ports of different
units. Therefore we use the internal signal generation to directly
reference an output of a unit.
- While in HDL both input and output ports need to be connected, in SMV
we just need to specify the inputs. Because there are no internal
signals in SMV, generating the ready signals requires us to look for the
users of each unit's output and search for the corresponding argument.
- The order of arguments is important in SMV, so we use `IOMap =
llvm::MapVector<Port, SmallVector<std::string>>;`, to keep track of the
order in which port-mappings are put into the IOMap.
- When a unit instantiates another, the ready signals of the parent unit
cannot be accessed by the child unit. For only this specific case in SMV
we define some internal signals. The other internal signals that are
defined are the outputs of the parent unit.
- Other differences are mostly syntactical and they didn't require any
major changes to the current infrastructure.