[Experimental] FTD algorithm (#299)
With respect to #177, this PR implements the core of the algorithm, as
reported in the original paper by Aya.
This PR is unfortunately very large, but it's somehow a mandatory step,
as it constitutes a full block necessary for the new conversion pass.
Also, notice that no conversion pass is PRed yet, so the flow cannot be
tested yet.

I mark this PR as draft as I'd like to receive suggestions in terms on
how to refactor to improve the clarity of the code. We are confident in
the functionalities, but we believe there is room for improvement in the
way the code is realised.