[Buffers][Refactoring] Remove Unused Internal Buffer Functions (#358)
This PR is part of #356. Specifically, it removes the unused functions
`addInternalBuffers` and `deductInternalBuffers`, along with their
associated data structures. Ideally, no functional changes are
introduced in this PR.

For more context, please refer to the discussions in #342. In short, no
unit will have a buffer placed at its input or output anymore. Instead,
such buffer placements are now explicitly defined as constraints in the
HandshakeSetBufferingProperties pass. This PR is the first step toward
that direction: removing the now-unnecessary code for deducting
input/output buffers.

The experimental results are available in
[Refactor_Results.pdf](https://github.com/user-attachments/files/19428609/SP25_InternalBuf.pdf).
Some differences are still observed in the outcomes.

Additionally, if we remove the logic that sets `minTrans` for Merge
units, the updated results can be found in
[NoMinTrans_Results.pdf](https://github.com/user-attachments/files/19428618/SP25_InternalBuf-2.pdf).
This modification generally leads to improved results.