[Speculation] Support Indirect Commit Units (#386)
This PR, together with #376 and #385, enables general execution of
if_convert by adding support for indirect commit units.

I decided to go ahead and implement this instead of writing a report
like "Support for synchronizer and indirect commit units is left for
future work, because ... In cases of ..., speculation doesn't work. A
possible solution would be to ..." I thought writing it would take more
time than just implementing it-which actually took me less than two
hours.

### Changes:

- **Commit unit placement beyond save-commit units**
- EDIT: Still stops the traversal at save-commits (or speculator (new)),
but...
- Start commit unit placement traversals from save-commit units
(`findCommitsReachableFromSCs`).
- Run `findCommitsBetweenBBs` after save-commit placement is finalized,
which now considers speculative edges reachable from save-commit units
but not from speculators.
- **Support commit and save-commit units on the same operand**
- Now we place `save-commit -> commit` (previously it was `commit ->
save-commit`). The order of function calls is reversed for correctness.
- **Update `routeCommitControl` to cover indirect commit units**
  - Now starts also from save-commit units.
- **Traversal still terminates at speculators or save-commit units to
avoid redundant branch duplication.**
- **Renaming of functions in PlacementFinder.cpp**

With this change, I believe we no longer need synchronizers.

If this direction is approved in today's meeting, I'll write a proper
design/documentation update for it.

---------

Co-authored-by: Emmet Murphy <45939330+murphe67@users.noreply.github.com>